Renamed iomediator00 to _iomediator. This module now only contains the IoMediator class and some helpers.
