+++ lib/autokey/iomediator/__init__.py
@@ -20,3 +20,4 @@ IoMediator public interface
 from ._windowgrabber import WindowGrabber
 from ._waiter import Waiter
 from ._keygrabber import Recorder, KeyGrabber
+from ._iomediator import IoMediator
